backend/requirements.txt
asgiref==3.7.2
black==22.10.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.8.0
flake8==5.0.4
identify==2.5.8
isort==5.10.1
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.10.1
Pillow==10.0.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.14
virtualenv==20.16.6
